ScoreTest: bad PMID for the main article.  When spidering through for pubmed it moves AUTHOR --> AUTHORSHIP --> PUBLICATION --> Then out to the rest of the nodes.  This broke down in the test because the URI with PMID for the authorship was not the same as the publication.
Match: kept most of the PHI, however moved from adding to the model every time to building a list of all of the statements, then adding the statements before moving to the next matchset.  Will clean up for memory usage this weekend.  Need full test on QA later this weekend.