Large number of updates to Peoplesoft-Biztalk example ingest:
This version removes the previous-harvest convention and direct VIVO
access in favor of a previous-data-state model and a proxy-vivo model.

Peopleexport.sh constructs the proxy VIVO of data relevant to the
incoming set, reducing the execution time of complex tools such as
scoring and selective diffing. The additions and subtractions computed
by the harvest are then applied against the real VIVO in the last
transfer.